A first implementation of the multiple rendering processes.
Messages are handled by the BApplication ( the only Looper in fact ), so the Tranquility class has got a BList of channel ( a Channel is a structure with a team_id (the renderBoy) and a pointer to the ProxyViewManager which manage this renderer )
A sad tab has also be added.
In my opinion, the name "ProxyViewManager" doesn't fit, it could even change.

There is still an issue when quitting the application using the "Quit Application" menu in the Deskbar.